not necessary

+++ bundles/net.tourbook.device.garmin.fit/src/net/tourbook/device/garmin/fit/listeners/MesgListener_Hr.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2005, 2019 Wolfgang Schramm and Contributors
+ * Copyright (C) 2005, 2023 Wolfgang Schramm and Contributors
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -119,7 +119,7 @@ public class MesgListener_Hr extends AbstractMesgListener implements HrMesgListe
 
          final long hr2TourTimeDiffMS = firstTourTimeMS - firstHrTimestampMS;
 
-         final long firstHrTimeSec = Float.valueOf(allEventTime[0]).longValue();
+         final long firstHrTimeSec = allEventTime[0].longValue();
          final long firstHrTimeMS = firstHrTimeSec * 1000;
 
          fitData.setTimeDiffMS(firstTourTimeMS - firstHrTimeMS - hr2TourTimeDiffMS);
